Zavrseno dodavanje
Uspjesno dodavanje, treba uvezati sa drugim i srediti da se iz istog moze editovati.

+++ src/org/unibl/etf/dto/HeightPriceRatioTableItem.java
@@ -1,13 +1,15 @@
 package org.unibl.etf.dto;
 
 import javafx.beans.property.DoubleProperty;
+import javafx.beans.property.IntegerProperty;
 import javafx.beans.property.SimpleDoubleProperty;
+import javafx.beans.property.SimpleIntegerProperty;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
 
 public class HeightPriceRatioTableItem {
 	private HeightPriceRatio ratio;
-	private DoubleProperty minHeight;
+	private IntegerProperty minHeight;
 	private StringProperty maxHeight;
 	private DoubleProperty price;
 
@@ -16,13 +18,13 @@ public class HeightPriceRatioTableItem {
 	}
 	public HeightPriceRatioTableItem(HeightPriceRatio ratio) {
 		this.ratio = ratio;
-		minHeight = new SimpleDoubleProperty(ratio.getMinHeight());
+		minHeight = new SimpleIntegerProperty(ratio.getMinHeight());
 		maxHeight = new SimpleStringProperty(ratio.getMaxHeight() == null ? "-" : ratio.getMaxHeight()+ "");
 		price = new SimpleDoubleProperty(ratio.getPrice());
 	}
 	
 	// properties
-	public DoubleProperty minHeightProperty() {
+	public IntegerProperty minHeightProperty() {
 		return minHeight;
 	}
 	public StringProperty maxHeightProperty() {
@@ -32,24 +34,24 @@ public class HeightPriceRatioTableItem {
 		return price;
 	}
 	// getteri
-	public Double getMinHeight() {
+	public Integer getMinHeight() {
 		return minHeight.get();
 	}
-	public Double getMaxHeight() {
+	public Integer getMaxHeight() {
 		return ratio.getMaxHeight();
 	}
 	public Double getPrice() {
 		return price.get();
 	}
 	// setteri
-	public void setMinHeight(Double minHeight) {
+	public void setMinHeight(Integer minHeight) {
 		ratio.setMinHeight(minHeight);
 		this.minHeight.set(minHeight);
 	}
 	public void setMaxHeight(String maxHeight) {
 		this.maxHeight.set(maxHeight);
 		if(!"-".equals(maxHeight)) {
-			ratio.setMaxHeight(Double.parseDouble(maxHeight));
+			ratio.setMaxHeight(Integer.parseInt(maxHeight));
 		}
 	}
 	public void setPrice(Double price) {
